Splits up and renames files to be more consistent
Also initializes history to be [0,0,0,0,0]
