+++ ng/ngculator/buttons/subtract/component.js
@@ -1,0 +1,15 @@
+import angular from 'angular';
+import template from '../button.html';
+
+const app = angular.module('ngculator');
+
+/**
+ * mbSubtractButton Compoent
+ * Subtraction button component
+ *
+ * @class mbSubtractButton
+ */
+app.component('mbSubtractButton', {
+  template: template,
+  controller: 'SubtractCtrl',
+});
