+++ ng/ngculator/buttons/subtract/controller.spec.js
