+++ ng/ngculator/frame/ngculator.html
@@ -1,8 +1,8 @@
-<mb-display></mb-display>
-<mb-input></mb-input>
+<nc-display></nc-display>
+<nc-input></nc-input>
 
-<mb-add-button></mb-add-button>
-<mb-subtract-button></mb-subtract-button>
+<nc-add-button></nc-add-button>
+<nc-subtract-button></nc-subtract-button>
 
 <br>
 <button ng-click="$ctrl.clear()">Clear</button>
