+++ ng/ngculator/modules/display/controller.js
@@ -1,5 +1,6 @@
+'use strict';
+
 import angular from 'angular';
-import template from './template.html';
 
 const app = angular.module('ngculator');
 
@@ -11,15 +12,4 @@ const app = angular.module('ngculator');
  */
 app.controller('DisplayCtrl', function (history) {
   this.history = history;
-});
-
-/**
- * mbDisplay Component
- * Displays the last 5 values
- *
- * @class DisplayCtrl
- */
-app.component('mbDisplay', {
-  template: template,
-  controller: 'DisplayCtrl',
 });
