+++ ng/ngculator/modules/input/component.js
@@ -1,0 +1,17 @@
+'use strict';
+
+import angular from 'angular';
+import template from './nc-input.html';
+
+const app = angular.module('ngculator');
+
+/**
+ * ncInput
+ * Input component for the calculator input.
+ *
+ * @class ncInput
+ */
+ app.component('ncInput', {
+   template: template,
+   controller: 'InputCtrl'
+ });
