+++ ng/ngculator/modules/input/controller.js
@@ -1,0 +1,16 @@
+'use strict';
+
+import angular from 'angular';
+
+const app = angular.module('ngculator');
+
+/**
+ * InputCtrl
+ * Controller for the MbInput directive
+ *
+ * @class InputCtrl
+ */
+
+ app.controller('InputCtrl', function (calcInput) {
+   this.input = calcInput;
+ });
